backend/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==11.2.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.4
google-generativeai==0.8.3
pypdf==5.3.0
python-dotenv==1.2.1
django-filter
openpyxl==3.1.5
xlrd==1.2.0
qrcode==7.4.2

# Async PDF generation (Gold Standard)
weasyprint==67.0
celery[redis]==5.4.0
django-anymail[mailgun]==12.0
